+++ src/renderer/components/uielements/alert/index.tsx
@@ -1,1 +1,3 @@
 export { Alert } from './Alert'
+export type { Color as AlertIconColor } from './AlertIcon.styles'
+export { AlertIcon } from './AlertIcon.styles'