+++ src/renderer/i18n/fr/pools.ts
@@ -11,9 +11,10 @@ const pools: PoolsMessages = {
   'pools.pending': 'En attente',
   'pools.available': 'Disponible',
   'pools.pooled': 'Mis en commun',
-  'pools.fundscap.reached': 'Limite de fonds atteinte',
-  'pools.fundscap.no': 'Pas de limite de fonds',
-  'pools.fundscap.error': "Erreur lors de l'obtention de la limite de fonds"
+  'pools.limit.info': 'Protocol limit reached! {pooled} pooled in pools, {bonded} bonded in nodes - FR',
+  'pools.incentivependulum.info': 'Incentive Pendulum: {percentage}%  - FR',
+  'pools.incentivependulum.tooltip': '{pooled} pooled in pools, {bonded} bonded in nodes - FR',
+  'pools.incentivependulum.error': 'Error while loading data of Incentive Pendulum - FR'
 }
 
 export default pools