+++ src/renderer/views/pools/ActivePools.tsx
@@ -15,14 +15,15 @@ import * as RxOp from 'rxjs/operators'
 
 import { Network } from '../../../shared/api/types'
 import { ManageButton } from '../../components/manageButton'
-import { FundsCap } from '../../components/pool'
+import { ProtocolLimit, IncentivePendulum } from '../../components/pool'
 import { Button } from '../../components/uielements/button'
 import { Table } from '../../components/uielements/table'
 import { useAppContext } from '../../contexts/AppContext'
 import { useMidgardContext } from '../../contexts/MidgardContext'
 import { ordBaseAmount, ordNumber } from '../../helpers/fp/ord'
 import * as PoolHelpers from '../../helpers/poolHelper'
-import { useFundsCap } from '../../hooks/useFundsCap'
+import { useIncentivePendulum } from '../../hooks/useIncentivePendulum'
+import { useProtocolLimit } from '../../hooks/useProtocolLimit'
 import * as poolsRoutes from '../../routes/pools'
 import { SwapRouteParams } from '../../routes/pools/swap'
 import { DEFAULT_NETWORK } from '../../services/const'
@@ -42,7 +43,9 @@ export const ActivePools: React.FC<PoolsComponentProps> = ({ haltedChains, mimir
   const network = useObservableState<Network>(network$, DEFAULT_NETWORK)
 
   const { service: midgardService } = useMidgardContext()
-  const { reload: reloadFundsCap, data: fundsCapRD } = useFundsCap()
+  const { reload: reloadLimit, data: limitRD } = useProtocolLimit()
+  const { data: incentivePendulumRD } = useIncentivePendulum()
+
   const {
     pools: { poolsState$, reloadPools, selectedPricePool$, poolsFilters$, setPoolsFilter }
   } = midgardService
@@ -67,8 +70,8 @@ export const ActivePools: React.FC<PoolsComponentProps> = ({ haltedChains, mimir
 
   const refreshHandler = useCallback(() => {
     reloadPools()
-    reloadFundsCap()
-  }, [reloadPools, reloadFundsCap])
+    reloadLimit()
+  }, [reloadPools, reloadLimit])
 
   const selectedPricePool = useObservableState(selectedPricePool$, PoolHelpers.RUNE_PRICE_POOL)
 
@@ -225,7 +228,8 @@ export const ActivePools: React.FC<PoolsComponentProps> = ({ haltedChains, mimir
               A.map(({ pool }) => pool.target)
             )}
           />
-          <FundsCap fundsCap={fundsCapRD} />
+          <ProtocolLimit limit={limitRD} />
+          <IncentivePendulum incentivePendulum={incentivePendulumRD} />
           <Table
             columns={columns}
             dataSource={FP.pipe(tableData, filterTableData(poolFilter))}
@@ -242,7 +246,16 @@ export const ActivePools: React.FC<PoolsComponentProps> = ({ haltedChains, mimir
         </>
       )
     },
-    [isDesktopView, desktopPoolsColumns, mobilePoolsColumns, poolFilter, setFilter, fundsCapRD, history]
+    [
+      isDesktopView,
+      desktopPoolsColumns,
+      mobilePoolsColumns,
+      poolFilter,
+      setFilter,
+      limitRD,
+      incentivePendulumRD,
+      history
+    ]
   )
 
   return (